requirements.txt
groq
streamlit
python-dotenv
PyMuPDF
gspread
google-auth
google-auth-oauthlib
google-api-python-client
requests
